requirements.txt
rdkit>=2022.09.5
py3Dmol>=2.0.1
MDAnalysis>=2.4.0
prolif>=1.1.0
numpy>=1.21.0
matplotlib>=3.5.0
PyYAML>=6.0
pandas>=1.5.0
requests>=2.28.0
progressbar2>=4.2.0